The impossible bet... solved?
Start by searching for the dollar bill in the box at index of the
person's place in line (minus one because arrays are 0 indexed).
If you didn't find it, try searching for the index of the dollar bill
you took out of the box. Do this within 50 tries for each person.